requirements.txt
boto3==1.9.96
botocore==1.12.96
certifi==2018.10.15
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.8
django-allauth==0.42.0
django-bootstrap-form==3.4
django-cors-headers==3.4.0
django-crispy-forms==1.9.1
django-filter==2.0
django-floppyforms==1.9.0
django-happenings==0.3.3
django-import-export==2.2.0
django-map-widgets==0.3.0
django-phonenumber-field==4.0.0
django-simple-email-confirmation==0.70
django-sslserver==0.22
django-verified-email-field==1.6.1
djangorestframework==3.11.0
djangorestframework-filters==0.11.1
django-heroku==0.3.1
django-storages==1.7.1
social-auth-app-django==4.0.0
social-auth-core==3.3.3
phonenumbers==8.12.6
docutils==0.14
gunicorn==19.9.0
idna==2.7
jmespath==0.9.3
Pillow==5.2.0
psycopg2==2.7.7
python-dateutil==2.8.0
pytz==2018.5
requests==2.19.1
s3transfer==0.2.0
six==1.12.0
urllib3==1.23
whitenoise==4.1.2
beautifulsoup4==4.7.1
numpy==1.16.4
Open-Tamil==0.97
urllib3==1.24.2